requirements.txt
pyserial
pynput
sounddevice
webrtcvad
openai
pyyaml
pyobjc-framework-Quartz; sys_platform == "darwin"
